+++ tencentcloud/vod/v20180717/client.go
@@ -8752,11 +8752,11 @@ func NewReviewAudioVideoResponse() (response *ReviewAudioVideoResponse) {
 }
 
 // ReviewAudioVideo
-// 对点播中的音视频媒体发起审核任务，智能检测视频画面、画面中的文字、语音中的文字出现的违规内容。
+// 对点播中的音视频媒体发起审核任务，智能检测视频画面、画面中的文字、语音中的文字、声音出现的违规内容。
 //
 // 
 //
-// 如使用事件通知，事件通知的类型为 [任务流状态变更](https://cloud.tencent.com/document/product/266/9636)。
+// 如使用事件通知，事件通知的类型为 [音视频审核完成](https://cloud.tencent.com/document/product/266/81258)。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION = "FailedOperation"
@@ -8773,11 +8773,11 @@ func (c *Client) ReviewAudioVideo(request *ReviewAudioVideoRequest) (response *R
 }
 
 // ReviewAudioVideo
-// 对点播中的音视频媒体发起审核任务，智能检测视频画面、画面中的文字、语音中的文字出现的违规内容。
+// 对点播中的音视频媒体发起审核任务，智能检测视频画面、画面中的文字、语音中的文字、声音出现的违规内容。
 //
 // 
 //
-// 如使用事件通知，事件通知的类型为 [任务流状态变更](https://cloud.tencent.com/document/product/266/9636)。
+// 如使用事件通知，事件通知的类型为 [音视频审核完成](https://cloud.tencent.com/document/product/266/81258)。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION = "FailedOperation"